chore: move list components into character folder
