+++ src/components/characters/search/SearchAndFilterBar.jsx
