+++ src/components/characters/ui/FavoriteButton.jsx
