+++ src/components/characters/ui/FavoriteButton.test.js
